+++ src/main/kotlin/de/moritzf/quota/idea/settings/QuotaSettingsConfigurable.kt
@@ -780,7 +780,7 @@ class QuotaSettingsConfigurable : Configurable {
                 account.setExtra(ProviderAccount.EXTRA_AZURE_ENDPOINT, panel.endpoint())
                 account.setExtra(ProviderAccount.EXTRA_AZURE_LOCATION, panel.locationId())
                 account.setExtra(ProviderAccount.EXTRA_AZURE_DEPLOYMENTS, panel.deploymentNames())
-                account.setExtra(ProviderAccount.EXTRA_AZURE_OCR_DEPLOYMENT, panel.ocrDeployment())
+                account.setExtra(ProviderAccount.EXTRA_AZURE_OCR_DEPLOYMENT, panel.ocrDeploymentForStorage())
             }
             QuotaProviderType.GITHUB ->
                 account.setExtra(